moved root to state.rs
