+++ crates/rpc/src/handlers/mod.rs
@@ -4,7 +4,6 @@ pub mod config;
 pub mod fork;
 pub mod genesis;
 pub mod randao;
-pub mod root;
 pub mod state;
 pub mod validator;
 pub mod version;
